stop reporting previous errors when all rerun
Issues #2040 and #2821 discuss an issue where previous failure counts
are displayed for tests stated as not rerun, even though all tests
have run successfully.  This is caused when a file remains that cannot
be traced to a specific tests triggers a rerunning of all tests.
When this happens prune does not remove the previous files with
failures as it is not run when specific files is empty (as in a rerun)
leading to incorrect reporting.  This fix sends a zero count for
previous failures if specificFiles array is empty.  Pruning will run
later to remove the old historic entries from the filesWithFailures array.